packagegroups: move Qt packages to right packagegroup
qtinterfaceframework packages should be added in the qt6 packagegroup and it's
required only for mingw toolchain.
DISTRO_FEATURE doesn't work as expected in nativesdk builds, so the wayland
additition hasn't done anything, we can remove it.

This fixes ci-sdk, so that it doesn't have any nativesdk-qt*-dev packages
installed, which interfere with the CI builds.

Pick-to: nanbield
Change-Id: I4315fac90fee9673fe8009e2c5efb4ad59e485c0
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>